Set stud_date from study via API again.
There's a bug in tomedo that doesn't allow to send the study_date for
the last visit.
Therefore we have to send the study_date along with the study field
again until this is fixed.
In tomedo the REST-configuration for the study field in
faxomat-Karteieintrag has to have this value:

        $[x UNT 1 bes inf NN NNNN NJNN _]$

This prefixed the study with the study date.